Change in User.in_sas to User.FAIL transition.
The gard on the transition is now '>=' instead of simply '>'.
This allow the models the go to the User.FAIL state if the controller is not on our side to take the transition back to Sas.idle at timeInSas == SAS_TIMER.
In the previous version the User had no choice, the Sas would always act at SAS_TIMER, making the resolution trivial.